WIP: Fix Cloud Functions in prod by moving all files to `test/screenshot/infra/`
