+++ test/screenshot/infra/.gcloudignore
