+++ test/screenshot/infra/deploy.sh
@@ -1,7 +1,5 @@
 #!/usr/bin/env bash
 
-gcloud datastore indexes create test/screenshot/shield/index.yaml
-
 cd $(dirname "$0")
 
 if [[ "$1" == '--local' ]]; then
@@ -9,7 +7,10 @@ if [[ "$1" == '--local' ]]; then
   functions start
   functions deploy screenshot-shield-svg --entry-point=screenshotShieldSvg --trigger-http
   functions deploy screenshot-shield-url --entry-point=screenshotShieldUrl --trigger-http
-else
+elif [[ "$1" == '--prod' ]]; then
+  gcloud datastore indexes create index.yaml
   gcloud functions deploy screenshot-shield-svg --entry-point=screenshotShieldSvg --runtime nodejs8 --trigger-http
   gcloud functions deploy screenshot-shield-url --entry-point=screenshotShieldUrl --runtime nodejs8 --trigger-http
+else
+  echo "Usage: $0 [--local|--prod]"
 fi
