+++ test/screenshot/infra/index.yaml
@@ -1,7 +1,7 @@
 #
 # IMPORTANT: If you modify this file, you'll also need to push it to GCP:
 #
-#     gcloud datastore indexes create test/screenshot/shield/index.yaml
+#     gcloud datastore indexes create test/screenshot/infra/index.yaml
 #
 # It may take several minutes for GCP to generate the indexes.
 #
